Lint: add basic ontology linting for structure and spelling
This commit starts to add additional linting for the OAP, in order to
validate the structure and content of the documents.  There are two
pieces to this commit:

 * Structural linting uses yamale to validate the OAP data against
   yaml schema for equipment, points, and tags files.  The schemas are
   based on the actual data present and so might not be exactly what
   the designers intended; but prevent indentation errors.
 * Spellchecking runs cspell on the text.  Identifiers and other words must be
   added to lint/ids.txt or the build will fail.
 * Two github workflows which run these linters on commit.

In the future, we should add additional check; eg to ensure each point
has a definition (see #2); but until that issue is closed the test
would fail anyways.